requirements.txt
flask==2.3.3
flask-mysqldb==1.0.1
flask-bcrypt==1.0.1
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
groq==0.4.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2